A great Time patch from John Hornbeck <hornbeck@gmail.com>:
A more complete Time diff.  This includes many of the instance methods for Time and a new primitive for usec.  This diff also includes some failing specs as I went ahead and added the specs for the rest of the class methods.  Also included is the constants for Time. 

Keep the good work coming John!