Change LRE protocol to use home context as the key
Rather than allocating an LRE everytime we want to do as end with a
block, we stash away the current context in a local that a future return
or break can read and use.

That return/break creates an LRE and sticks the home's context inside
it. The wrapper code around a send_with_block then checks if the
exception is an LRE AND it's for the current context. If so, it handles
it. Otherwise, it raises in on up.