Melbourne visitor calls processor methods with all children.
This is only step one. The Node.from method is being used while
all the compiler specs are made to pass. Then all the nodes will
be rewritten to use .new instead of .from (i.e. once the existing
compiler can be piecewise updated without all sorts of confusion).

The specs can be run as follows:

  bin/mspec --ast --compiler spec/compiler/lit_spec.rb

Much remains to be done, principally:

1. create accessors for data that was processed from the sexp,
   including looking up or down the tree for data (eg #get, #set).
2. redo some of the sexp normalization via node creation/processing
   instead (eg method call arguments).