Outline of new compiler infrastructure.
Q. What is a compiler?
A. It's a transformation.

Q. What's it do?
A. Transforms stuff.

Q. What's the goal of these here codes?
A. Create an API that simplifies running the compiler in various
   contexts like compiling eval strings, loading Ruby code scripts,
   compiling files with a command line compiler script, etc.

Q. And?
A. And this is done by making the stages of compilation explicit
   by exposing them as objects arranged into a pipeline.

Q. The compiler's got no clothes!
A. Huh?

Q. Can't you die of exposure.
A. Yes.

Q. Go on.
A. The compiler *is* a pipeline. Each compiler instance takes input,
   processes it through one or more stages, and finally returns the
   result.

Q. And a stage has actors?
A. No. Each stage performs one well-contained step in the process.
   For example, the parser stage takes Ruby source code and produces
   an abstract syntax tree (AST).

Q. Isn't that a lot like this:

    pipeline.inject(source) { |form, stage| stage.process form }

A. Yes.

Q. So why have a compiler at all? Why not just throw some stages in
   an array and map them?
A. You could.

Q. Ok, but instead you've got this linked structure. Why?
A. To simplify creating different kinds of compilers. For example,
   one that takes a string and produces symbolic bytecode. The
   compiler tests do this. Or one that takes an AST and produces
   a compiled method.

Q. Is that all?
A. No. It also simplifies adding new stages. Each stage knows which
   stage follows it. Well, it thinks it knows. But this can easily
   be altered to insert a new stage.

Q. What else does a stage do, besides this linking stuff?
A. A stage also knows how to invoke the processor for that stage.
   Essentially, it receives input, invokes the processor, and
   passes the data produced to the next stage. For example, the
   parser stage takes the source code, invokes the Ruby parser,
   and passes the AST to the next stage.

Q. Why have stages at all? Couldn't you invoke the parser directly?
A. Yes. But the stage encapsulates the logic of invoking the
   processor and simplifies using different processors. For example,
   the bytecode generation stage uses a Generator normally but uses
   a TestGenerator in the compiler specs.

Q. And why would I care about inserting a stage?
A. You could, for instance, have a stage that takes input in XML
   format and produces an AST. Or you could have a different writer
   stage that produces an encrypted compiled file.

Q. Isn't that excessive?
A. Do you think when the men went to the moon, designing rungs on
   the lunar lander's ladder was excessive? Couldn't the guys just
   jump in and out of the lander in the moon's 1/6th earth gravity?
   Also, why did none of the astronauts do back flips or have a
   WWE wrestling competition?

Q. So we have compilers, stages, and processors of some sort?
A. Yes. The compiler is a pipeline of stages. The stages have an
   interface with the compiler for creating stages, an interface
   with the preceeding and succeeding stage, and an interface
   with the processor(s) for that stage.

Q. Isn't there a better way to do this?
A. No. Well, maybe.