Added AST transforms.
The transforms were previously implemented as compiler plugins.
Here, the functionality is split in two parts: 1) recognizing
a form that will be transformed and returning a node that will
emit different bytecode, 2) emitting the bytecode.

The recognition step is essentially stateless and is implemented
as class methods on each transform class. For now, only the CALL
parse tree node is processed for possible transforms.

The transform classes register themselves under a category. For
now, the categories are :default and :kernel. The catergory of
transforms to be applied is selected for each compiler instance.
Instead of applying all the transforms in a category, it is also
possible to apply only a selected set of transforms.