Introduce State to Generator.
This replaces the various map_xxx methods use previously. The issue
with the previous scheme is that it creates many unnecesary Arrays
and visits nodes that care nothing about the state being set. This
is obvious from the various empty in_xxx methods on Node.

This scheme implements various state stacks in AST::State. The state
is pushed before processing children and popped after processing the
children. Stacks are needed for nested AST structures that would set
the same state (i.e. a simple flag would not work because a nested
tree would unset the flag).