Requiem for an era: o/~ oh compiler, your sexp expressions... o/~
Let's take a moment to reflect an all this compiler has done for us.
It took a lot of work and it took us a long way. It was my stalwart
companion as I learned all the nooks and crannies of Ruby (ok, I'm
still learning). And now (gulp) I'll have to synthesize the correct
bytecode from scratch!

Thank you compiler. We won't miss you, especially not your sexps,
but we appreciate all you've done for us.