Fixed next running ensures.
Note that this modifies slightly emitting a loop body so that discarding
the value computed by the loop and checking interrupts is done in an
epilogue to the loop body. This epilogue is jumped to by e.g. the next
statement and flowed to by normal loop scenarios. No computation should
be performed in the epilogue but this naturally locates things like
check_interrupts and keeps them with the loop rather than spread around
like in the code for the next statement.