Initial commit of Channel9 VM/language
The idea behind this is a complete VM environment built around
continuation passing. The plan is to use this as a platform for an event
driven framework.

It's not exactly stackless, but it does use very small stacks, with
lexical closures and continuation linking as the means by which 'stack'
traversal occurs.