Implemented the first primitive: Message.
A Message object is one that is used to bundle together a message to be
sent to a receiver. It can include positional arguments, named
arguments, and a name. The information can be used by an object to
dispatch a method call, and unpack arguments in the stack layout needed.