VM Environment now starts with Context#channel_send
Previously you set a context into the environment and then ran it. This
is kind of redundant and error prone, so now you set up your environment
and context, and then in order to actually start the VM you just do a
channel_send on the context and it goes (if the VM isn't already).