Implemented 'system' arguments to messages.
In order for extra information to be conveyed to a message that may
otherwise be difficult to pass in in a uniform way, you can now pass
system arguments into a method. It's essentially a parallel argument
stack to the argument list the application itself sees, and can be used
for flags or blocks or anything else of that nature. Channel9.rb is
using it to pass the yielded block in.