Made it possible to create amalgamated strings.
The string_new instruction allows for the creation of a new string
object from several off the stack. This also marks the first instruction
that calls back into the VM to do work, and I'm not sure the current
incarnation of that functionality isn't a little brittle. I guess we'll
see, though.