Generalized unwind semantics.
Made it so that there's a currently registered unwind handler, and
established a protocol for sending unwind messages through it.
Essentially, an unwind handler is registered by sending a message to a
special channel and storing the old value. On reaching the unwind
handler, it re-registers the previous handler and does its processing.

Unwind handlers take a message that indicates what kind of unwind it is.
Examples would include :long_return for a return through ensure or
iterator blocks, :exception for exceptions, :throw for a throw, :break
for a break from an iterator, etc. As of this commit, only :long_return
is implemented.