Do less duping
Made it so that making a channel doesn't do any major duping until the
channel is actually called into. Also made it so that the return channel
of a method is just the current context without a new stack. This has
potential problems down the road for situations where a method might be
returned from twice (ie. some kind of fork), but I'll cross that bridge
when I come to it.