Added case statements. (Along with many refactorings)
As prelude to exceptions, implemented the case/when construct. In the
process, properly implemented operator === (at least for a subset of
classes) and fixed a whole bunch of wrong things about the object model.

In particular, singleton classes now (should) have correct inheritence
(a singleton class inherits from the singleton class of its superclass,
and the singleton class of class inherits from itself), and lookup
proceeds through the singleton hierarchy before going through the normal
class hierarchy.

I'd just like to go on record as saying that the singleton class of
Class inheriting from itself is horribly obnoxious and annoying in how
it doesn't correspond to the normal class inheritence tree. It seems
this was fixed in ruby 1.9 (and rbx, for that matter), but for now I'm
targetting 1.8.7. In ruby 1.9, it goes
Class:Class->Class:Module->Class:Object->Class:BasicObject->Class.