Implemented naive versions of basic string evals.
Note that there are some seriously incorrect things about these
implementations of eval. Namely, they do not support local variable
binding at *all*, and class_eval is implemented as just instance_eval on
the class. This is incorrect because normally instance_eval should act
on the metaclass of the class instead of the class itself, but I'm not
sure how to actually implement that at this stage.