Big change to constant lookup to make it dynamic.
Previous implementation of constant lookup was incorrect in that it
assumed that there was a constant-time scope lookup. There is not. Scope
is determined at run-time. So the scope is now defined in the frame-var
const-self, which is a tuple-tree of module/class objects to look in. It
gets picked up by all class/method/module definitions done in its scope,
and added to if needed.