Hook up some of the GC plumbing.
There is now a mechanism for two actions related to the GC:
* scan: looks through the child objects of the object being called on
 and reallocates any immediate children.
* reallocate: build a new copy of the same object. References within the
 object are left alone.

This should allow for a process by which top level items are scanned,
which will cause their child objects to be reallocated. Then the new
heap can be scanned and objects in it can be scanned to cause their
child references to reallocate. And so on until no new heap entries are
created, at which point the collection is complete.

There are probably gaps in here, but it should help get the whole thing
started.