Split RunnableContext into runnable and running.
This eliminates an annoying need to manipulate the runnablecontext
post-collection to make the stack pointer point into the right place.

This is also a good idea in general, as the checks to see whether the
context was already running were complicating the code. Now the context
is exactly what it says it is.