Reorganized loading a bit.
You can now load a .c9l file that lists other things to load.
You can (theoretically) load a .so that can do stuff to the environment.
ARGV is passed in as a special variable on the environment so they
can do something about loading an actual script.

The idea is that this will form the core of booting up a language's
environment. A .c9l will list the steps needed to load, say, Ruby
and then the last file listed should take care of parsing out
argv and loading the actual ruby script.