Made it mostly load and run things without ruby.
This is a pretty big change to the whole setup. It now loads as a
c9 list file and a .so that does the most fundamental work, following into
loading all the rest of the code necessary to boot the system, and then
finally loading whatever was passed in as the program name to bin/c9
as the ruby code to run.

It doesn't entirely work, and it definitely doesn't compile ruby code
itself, but it's probably good enough to get a start on the rest and to
do some of the benchmarking stuff Timo wanted to do. It also leaves it
pretty much free of ruby dependencies at runtime.