Made environment loading its own main entry point.
You can now run a program with bin/c9 that has a valid c9-env setup
by just specifying that program on the command line:
 bin/c9 test.rb
will try to load the environment initialization scripts in:
 {path-to-libc9.so}/c9-env/rb/rb.c9l
and will pass in an argv tuple as a special parameter. The
environment scripts can then use that to run the actual program.