Reorganized tracing a bunch. Can now specify as part of cmake setup.
* A fairly complex macro replacement mechanism is used to determine
whether a given trace call should go through to output. Not as
nice as before, but it works.
* You can now set trace levels per-facility instead of just one level
for all subscribed traces.
* CMake can properly set trace levels for everything, and you can
use cmake-gui to set them in a nice way (or as cmdline -D flags
for a more oldschool way).
* bin/c9 disables tracing at start unless you specify -TT,
and it passes in to the environment that it should trace once loaded
if you pass it -T.